+++ src/main/java/dev/ftb/mods/mecrh/datagen/ModLangProvider.java
@@ -23,6 +23,7 @@ public class ModLangProvider extends LanguageProvider {
         add("item.mecrh.music_disc_chicken_of_chaos.desc", "FTB Team - Chicken of Chaos");
 
         add("mecrh.message.wrong_forcefield_item", "Invalid weapon for the forcefield!");
+        add("mecrh.message.chicken_despawn_warning", "No players in arena! Chicken despawns in %s seconds!");
 
         add("mecrh.subtitle.forcefield_on", "Forcefield activates");
         add("mecrh.subtitle.forcefield_off", "Forcefield deactivates");
@@ -31,5 +32,7 @@ public class ModLangProvider extends LanguageProvider {
         add("mecrh.subtitle.laser_start", "Laser powers up");
         add("mecrh.subtitle.laser_loop", "Laser firing");
         add("mecrh.subtitle.laser_end", "Laser powers down");
+        add("mecrh.subtitle.clear_warn", "Chicken shrieks");
+        add("mecrh.subtitle.chaos_hurt", "Chicken hurt");
     }
 }