+++ src/main/java/dev/ftb/mods/mecrh/registry/ModSounds.java
@@ -17,24 +17,11 @@ public class ModSounds {
     public static final DeferredHolder<SoundEvent, SoundEvent> LASER_LOOP = register("laser_loop");
     public static final DeferredHolder<SoundEvent, SoundEvent> LASER_END = register("laser_end");
     public static final DeferredHolder<SoundEvent, SoundEvent> CHARGE_START = register("charge_start");
-    public static final DeferredHolder<SoundEvent, SoundEvent> CRACK = register("crack");
-    public static final DeferredHolder<SoundEvent, SoundEvent> CRACK_OPEN = register("crack_open");
-    public static final DeferredHolder<SoundEvent, SoundEvent> EGG_AMBIENCE = register("egg_ambience");
     public static final DeferredHolder<SoundEvent, SoundEvent> FF_ON = register("ff_on");
     public static final DeferredHolder<SoundEvent, SoundEvent> FF_OFF = register("ff_off");
-    public static final DeferredHolder<SoundEvent, SoundEvent> CHAOS_MUSIC_INTRO = register("chaos_music_intro");
-    public static final DeferredHolder<SoundEvent, SoundEvent> CHAOS_MUSIC_OLD = register("chaos_music");
     public static final DeferredHolder<SoundEvent, SoundEvent> CHAOS_MUSIC = register("chicken_of_chaos");
-    public static final DeferredHolder<SoundEvent, SoundEvent> CHAOS_INTRO_1 = register("chaos_intro_1");
-    public static final DeferredHolder<SoundEvent, SoundEvent> CHAOS_INTRO_2 = register("chaos_intro_2");
-    public static final DeferredHolder<SoundEvent, SoundEvent> CHAOS_INTRO_3 = register("chaos_intro_3");
-    public static final DeferredHolder<SoundEvent, SoundEvent> CHAOS_INTRO_4 = register("chaos_intro_4");
     public static final DeferredHolder<SoundEvent, SoundEvent> CHAOS_HURT = register("chaos_hurt");
-    public static final DeferredHolder<SoundEvent, SoundEvent> CHAOS_DEATH = register("chaos_death");
-    public static final DeferredHolder<SoundEvent, SoundEvent> CHAOS_DEATH_SOUND = register("chaos_death_sound");
-    public static final DeferredHolder<SoundEvent, SoundEvent> CHAOS_CLEAR = register("chaos_clear");
     public static final DeferredHolder<SoundEvent, SoundEvent> CLEAR_WARN = register("clear_warn");
-    public static final DeferredHolder<SoundEvent, SoundEvent> CHAOS_SPINNY = register("chaos_spinny");
     public static final DeferredHolder<SoundEvent, SoundEvent> CHICKEN_SPIN = register("ender_death");
 
     private static DeferredHolder<SoundEvent,SoundEvent> register(String name) {