+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [21.1.4]
+
+### Added
+- Players near the Ender Chicken (but outside the arena radius) are now warned every 5 seconds if despawn is imminent due to lack of players in the arena
+
+### Fixed
+- Fixed the Ender Chicken leaving an uninteractable zombie rider behind if it despawned due to no players in arena
+
 ## [21.1.3]
 
 ### Changed